Thus the epic saga of debugging the AMR-induced instabilities in the moment-EM solver continues. I've removed the AMR version of the 10-moment parallel firehose instability test (because it didn't work), and have replaced it with an AMR version of the 10-moment generalized Brio-Wu Riemann problem test (which doesn't work either, but is somehow more diagnostically useful). I have tracked the problem down to the gkyl_moment_em_coupling_implicit_advance call helpfully replacing the entire domain with NaNs on the second time-step (but curiously never the first). Strange, and requires further analysis. I also fixed a minor typo/bug in the patch-structured AMR update function when threads are switched off (which I happened to catch incidentally while debugging this).
